#30: Moved migrations to separate folders for more clear project structure
