+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/DependenciesTests/DependencyCodeMigrations/CodeMigration_5_0_NotOkDependencies.cs
@@ -3,7 +3,7 @@ using System.Data.Common;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Curiosity.Migrations.IntegrationTests.DependencyCodeMigrations;
+namespace Curiosity.Migrations.IntegrationTests.DependenciesTests.DependencyCodeMigrations;
 
 public class CodeMigration_5_0_NotOkDependencies : CodeMigration, IDependencyMigration
 {
