+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/DependenciesTests/DependencyScriptMigrations/1.0-correct_script.sql
