+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/DependenciesTests/DependencyScriptMigrations/4.0-script_with_ok_dependencies.sql
