+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/DependenciesTests/DependencyScriptMigrations/7.0-script_with_not_ok_dependencies.sql
