+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/DependenciesTests/DependencyTests.cs
@@ -1,31 +1,31 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
 using System.Threading.Tasks;
-using Curiosity.Migrations.IntegrationTests.TransactionCodeMigrations;
+using Curiosity.Migrations.IntegrationTests.DependenciesTests.DependencyCodeMigrations;
 using Curiosity.Migrations.PostgreSQL;
 using Xunit;
 
-namespace Curiosity.Migrations.IntegrationTests;
+namespace Curiosity.Migrations.IntegrationTests.DependenciesTests;
 
-public class TransactionTests
+public class DependencyTests
 {
     [Fact]
-    public async Task Migrate_AllScriptOk_NoRollback()
+    public async Task Migrate_Script_OkDependencies()
     {
         var config = ConfigProvider.GetConfig();
-        var connectionString = String.Format(config.ConnectionStringMask, "test_ok");
+        var connectionString = String.Format(config.ConnectionStringMask, "test_script_code_ok");
 
 
         var builder = new MigrationEngineBuilder();
-        builder.UseCodeMigrations().FromAssembly<ITransactionMigration>(Assembly.GetExecutingAssembly());
-        builder.UseScriptMigrations().FromDirectory(Path.Combine(Directory.GetCurrentDirectory(), "TransactionScriptMigrations"));
+        builder.UseCodeMigrations().FromAssembly<IDependencyMigration>(Assembly.GetExecutingAssembly());
+        builder.UseScriptMigrations().FromDirectory(Path.Combine(Directory.GetCurrentDirectory(), "DependenciesTests/DependencyScriptMigrations"));
         builder.ConfigureForPostgreSql(connectionString);
 
         builder.UseUpgradeMigrationPolicy(MigrationPolicy.AllAllowed);
         builder.UseDowngradeMigrationPolicy(MigrationPolicy.AllAllowed);
-        builder.SetUpTargetVersion(new MigrationVersion(3));
+        builder.SetUpTargetVersion(new MigrationVersion(4));
 
         var migrator = builder.Build();
 
@@ -40,22 +40,23 @@ public class TransactionTests
         {
             new(1),
             new(2),
-            new(3)
+            new(3),
+            new(4)
         };
         Assert.True(result.IsSuccessfully);
         Assert.Equal(expectedAppliedMigrations, actualAppliedMigrations);
     }
-
+    
     [Fact]
-    public async Task Migrate_AllScriptOk_SwitchedOffTransaction()
+    public async Task Migrate_Script_CodeNotOkDependencies()
     {
         var config = ConfigProvider.GetConfig();
-        var connectionString = String.Format(config.ConnectionStringMask, "test_without_transactions");
+        var connectionString = String.Format(config.ConnectionStringMask, "test_code_not_ok");
 
 
         var builder = new MigrationEngineBuilder();
-        builder.UseCodeMigrations().FromAssembly<ITransactionMigration>(Assembly.GetExecutingAssembly());
-        builder.UseScriptMigrations().FromDirectory(Path.Combine(Directory.GetCurrentDirectory(), "TransactionScriptMigrations"));
+        builder.UseCodeMigrations().FromAssembly<IDependencyMigration>(Assembly.GetExecutingAssembly());
+        builder.UseScriptMigrations().FromDirectory(Path.Combine(Directory.GetCurrentDirectory(), "DependenciesTests/DependencyScriptMigrations"));
         builder.ConfigureForPostgreSql(connectionString);
 
         builder.UseUpgradeMigrationPolicy(MigrationPolicy.AllAllowed);
@@ -64,7 +65,7 @@ public class TransactionTests
 
         var migrator = builder.Build();
 
-        await migrator.UpgradeDatabaseAsync();
+        var result = await migrator.UpgradeDatabaseAsync();
 
         var migrationProvider = new PostgresMigrationConnection(new PostgresMigrationConnectionOptions(connectionString));
         await migrationProvider.OpenConnectionAsync();
@@ -76,41 +77,31 @@ public class TransactionTests
             new(1),
             new(2),
             new(3),
-            new(4),
-            new(5)
+            new(4)
         };
+        Assert.True(result.ErrorCode == MigrationErrorCode.MigratingError);
         Assert.Equal(expectedAppliedMigrations, actualAppliedMigrations);
     }
-
+    
     [Fact]
-    public async Task Migrate_AllScriptOk_Rollback()
+    public async Task Migrate_Script_ScriptNotOkDependencies()
     {
         var config = ConfigProvider.GetConfig();
-        var connectionString = String.Format(config.ConnectionStringMask, "test_rollback");
+        var connectionString = String.Format(config.ConnectionStringMask, "test_script_not_ok");
 
 
         var builder = new MigrationEngineBuilder();
-        builder.UseCodeMigrations().FromAssembly<ITransactionMigration>(Assembly.GetExecutingAssembly());
-        builder.UseScriptMigrations().FromDirectory(Path.Combine(Directory.GetCurrentDirectory(), "TransactionScriptMigrations"));
+        builder.UseCodeMigrations().FromAssembly<IDependencyMigration>(Assembly.GetExecutingAssembly());
+        builder.UseScriptMigrations().FromDirectory(Path.Combine(Directory.GetCurrentDirectory(), "DependenciesTests/DependencyScriptMigrations"));
         builder.ConfigureForPostgreSql(connectionString);
 
-        builder.UseUpgradeMigrationPolicy(MigrationPolicy.AllAllowed);
+        builder.UseUpgradeMigrationPolicy(MigrationPolicy.ShortRunningAllowed);
         builder.UseDowngradeMigrationPolicy(MigrationPolicy.AllAllowed);
-        builder.SetUpTargetVersion(new MigrationVersion(6));
+        builder.SetUpTargetVersion(new MigrationVersion(7));
 
         var migrator = builder.Build();
 
-        try
-        {
-            await migrator.UpgradeDatabaseAsync();
-
-            // last migration is incorrect, can not go here
-            Assert.False(true);
-        }
-        catch
-        {
-            // ignored
-        }
+        var result = await migrator.UpgradeDatabaseAsync();
 
         var migrationProvider = new PostgresMigrationConnection(new PostgresMigrationConnectionOptions(connectionString));
         await migrationProvider.OpenConnectionAsync();
@@ -122,10 +113,9 @@ public class TransactionTests
             new(1),
             new(2),
             new(3),
-            new(4),
-            new(5)
+            new(4)
         };
-
+        Assert.True(result.ErrorCode == MigrationErrorCode.MigratingError);
         Assert.Equal(expectedAppliedMigrations, actualAppliedMigrations);
     }
 }
