--- tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/DependencyCodeMigrations/IDependencyMigration.cs
@@ -1,6 +1,0 @@
-﻿namespace Curiosity.Migrations.IntegrationTests.DependencyCodeMigrations;
-
-public interface IDependencyMigration
-{
-    
-}
