--- tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/TransactionCodeMigrations/ITransactionMigration.cs
@@ -1,6 +1,0 @@
-﻿namespace Curiosity.Migrations.IntegrationTests.TransactionCodeMigrations;
-
-public interface ITransactionMigration
-{
-    
-}
