+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/TransactionsTests/TransactionCodeMigrations/CodeMigration_5_0 _TransactionOff.cs
@@ -5,7 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 
-namespace Curiosity.Migrations.IntegrationTests.TransactionCodeMigrations;
+namespace Curiosity.Migrations.IntegrationTests.TransactionsTests.TransactionCodeMigrations;
 
 public class CodeMigration_5_0 : CodeMigration, ITransactionMigration
 {
