+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/TransactionsTests/TransactionScriptMigrations/1.0-correct_script.sql
