+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/TransactionsTests/TransactionScriptMigrations/4.0-script_without_transction.sql
