+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/TransactionsTests/TransactionScriptMigrations/6.0-incorrect_script.sql
