+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/TransactionsTests/TransactionTests.cs
@@ -1,31 +1,31 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
 using System.Threading.Tasks;
-using Curiosity.Migrations.IntegrationTests.DependencyCodeMigrations;
+using Curiosity.Migrations.IntegrationTests.TransactionsTests.TransactionCodeMigrations;
 using Curiosity.Migrations.PostgreSQL;
 using Xunit;
 
-namespace Curiosity.Migrations.IntegrationTests;
+namespace Curiosity.Migrations.IntegrationTests.TransactionsTests;
 
-public class DependencyTests
+public class TransactionTests
 {
     [Fact]
-    public async Task Migrate_Script_OkDependencies()
+    public async Task Migrate_AllScriptOk_NoRollback()
     {
         var config = ConfigProvider.GetConfig();
-        var connectionString = String.Format(config.ConnectionStringMask, "test_script_code_ok");
+        var connectionString = String.Format(config.ConnectionStringMask, "test_ok");
 
 
         var builder = new MigrationEngineBuilder();
-        builder.UseCodeMigrations().FromAssembly<IDependencyMigration>(Assembly.GetExecutingAssembly());
-        builder.UseScriptMigrations().FromDirectory(Path.Combine(Directory.GetCurrentDirectory(), "DependencyScriptMigrations"));
+        builder.UseCodeMigrations().FromAssembly<ITransactionMigration>(Assembly.GetExecutingAssembly());
+        builder.UseScriptMigrations().FromDirectory(Path.Combine(Directory.GetCurrentDirectory(), "TransactionsTests/TransactionScriptMigrations"));
         builder.ConfigureForPostgreSql(connectionString);
 
         builder.UseUpgradeMigrationPolicy(MigrationPolicy.AllAllowed);
         builder.UseDowngradeMigrationPolicy(MigrationPolicy.AllAllowed);
-        builder.SetUpTargetVersion(new MigrationVersion(4));
+        builder.SetUpTargetVersion(new MigrationVersion(3));
 
         var migrator = builder.Build();
 
@@ -40,23 +40,22 @@ public class DependencyTests
         {
             new(1),
             new(2),
-            new(3),
-            new(4)
+            new(3)
         };
         Assert.True(result.IsSuccessfully);
         Assert.Equal(expectedAppliedMigrations, actualAppliedMigrations);
     }
-    
+
     [Fact]
-    public async Task Migrate_Script_CodeNotOkDependencies()
+    public async Task Migrate_AllScriptOk_SwitchedOffTransaction()
     {
         var config = ConfigProvider.GetConfig();
-        var connectionString = String.Format(config.ConnectionStringMask, "test_code_not_ok");
+        var connectionString = String.Format(config.ConnectionStringMask, "test_without_transactions");
 
 
         var builder = new MigrationEngineBuilder();
-        builder.UseCodeMigrations().FromAssembly<IDependencyMigration>(Assembly.GetExecutingAssembly());
-        builder.UseScriptMigrations().FromDirectory(Path.Combine(Directory.GetCurrentDirectory(), "DependencyScriptMigrations"));
+        builder.UseCodeMigrations().FromAssembly<ITransactionMigration>(Assembly.GetExecutingAssembly());
+        builder.UseScriptMigrations().FromDirectory(Path.Combine(Directory.GetCurrentDirectory(), "TransactionsTests/TransactionScriptMigrations"));
         builder.ConfigureForPostgreSql(connectionString);
 
         builder.UseUpgradeMigrationPolicy(MigrationPolicy.AllAllowed);
@@ -65,7 +64,7 @@ public class DependencyTests
 
         var migrator = builder.Build();
 
-        var result = await migrator.UpgradeDatabaseAsync();
+        await migrator.UpgradeDatabaseAsync();
 
         var migrationProvider = new PostgresMigrationConnection(new PostgresMigrationConnectionOptions(connectionString));
         await migrationProvider.OpenConnectionAsync();
@@ -77,31 +76,41 @@ public class DependencyTests
             new(1),
             new(2),
             new(3),
-            new(4)
+            new(4),
+            new(5)
         };
-        Assert.True(result.ErrorCode == MigrationErrorCode.MigratingError);
         Assert.Equal(expectedAppliedMigrations, actualAppliedMigrations);
     }
-    
+
     [Fact]
-    public async Task Migrate_Script_ScriptNotOkDependencies()
+    public async Task Migrate_AllScriptOk_Rollback()
     {
         var config = ConfigProvider.GetConfig();
-        var connectionString = String.Format(config.ConnectionStringMask, "test_script_not_ok");
+        var connectionString = String.Format(config.ConnectionStringMask, "test_rollback");
 
 
         var builder = new MigrationEngineBuilder();
-        builder.UseCodeMigrations().FromAssembly<IDependencyMigration>(Assembly.GetExecutingAssembly());
-        builder.UseScriptMigrations().FromDirectory(Path.Combine(Directory.GetCurrentDirectory(), "DependencyScriptMigrations"));
+        builder.UseCodeMigrations().FromAssembly<ITransactionMigration>(Assembly.GetExecutingAssembly());
+        builder.UseScriptMigrations().FromDirectory(Path.Combine(Directory.GetCurrentDirectory(), "TransactionsTests/TransactionScriptMigrations"));
         builder.ConfigureForPostgreSql(connectionString);
 
-        builder.UseUpgradeMigrationPolicy(MigrationPolicy.ShortRunningAllowed);
+        builder.UseUpgradeMigrationPolicy(MigrationPolicy.AllAllowed);
         builder.UseDowngradeMigrationPolicy(MigrationPolicy.AllAllowed);
-        builder.SetUpTargetVersion(new MigrationVersion(7));
+        builder.SetUpTargetVersion(new MigrationVersion(6));
 
         var migrator = builder.Build();
 
-        var result = await migrator.UpgradeDatabaseAsync();
+        try
+        {
+            await migrator.UpgradeDatabaseAsync();
+
+            // last migration is incorrect, can not go here
+            Assert.False(true);
+        }
+        catch
+        {
+            // ignored
+        }
 
         var migrationProvider = new PostgresMigrationConnection(new PostgresMigrationConnectionOptions(connectionString));
         await migrationProvider.OpenConnectionAsync();
@@ -113,9 +122,10 @@ public class DependencyTests
             new(1),
             new(2),
             new(3),
-            new(4)
+            new(4),
+            new(5)
         };
-        Assert.True(result.ErrorCode == MigrationErrorCode.MigratingError);
+
         Assert.Equal(expectedAppliedMigrations, actualAppliedMigrations);
     }
 }
